Move alertmanager api silence code to separate files unchanged
